+++ src/Core/Dfe.Complete.Domain/Constants/UserPolicyConstants.cs
@@ -13,4 +13,5 @@ public static class UserPolicyConstants
     public const string CanViewAllProjectsExports = "CanViewAllProjectsExports";
     public const string HasGlobalNavigation = "HasGlobalNavigation";
     public const string CanEditAddedByUser = "CanEditAddedByUser";
+    public const string ManageLocalAuthorities = "ManageLocalAuthorities";
 }