+++ src/Core/Dfe.Complete.Infrastructure/InfrastructureServiceCollectionExtensions.cs
@@ -36,6 +36,7 @@ namespace Dfe.Complete.Infrastructure
             services.AddScoped<IConversionCsvQueryService, ConversionCsvQueryService>();
             services.AddScoped<IProjectReadRepository, ProjectReadRepository>();
             services.AddScoped<IReadUserRepository,ReadUserRepository>();
+            services.AddScoped<ILocalAuthoritiesQueryService, LocalAuthoritiesQueryService>();
 
             // Authentication
             //services.AddCustomAuthorization(config);