+++ src/Tests/Dfe.Complete.CypressTests/cypress/pages/validationComponent.ts
@@ -26,6 +26,19 @@ class ValidationComponent {
 
         return this;
     }
+
+    public hasLinkedValidationErrorForField(fieldId: string, message: string): this {
+        cy.get(`.govuk-error-summary a[href="#${fieldId}"]`)
+            .contains(message)
+            .should("exist")
+            .invoke("attr", "href")
+            .then((href: string | undefined) => {
+                cy.get(href as string).should("exist");
+                cy.get((href as string) + "-error").should("contain.text", message);
+            });
+
+        return this;
+    }
 }
 
 const validationComponent = new ValidationComponent();