+++ CHANGELOG.md
@@ -22,7 +22,11 @@ Security in case of vulnerabilities.
 
 ## Unreleased  
 
-### Added  
+### Added
+- New route /projects/service-support/without-academy-urn
+- Service support Conversion URNs (/projects/service-support/without-academy-urn)
+- New route /projects/{projectId}/academy-urn
+- Service support Create Academy URN (/projects/{projectId}/academy-urn) 
 - Added `Statistics` page
 
 ### Changed